+++ Tortuga.Chain/Tortuga.Chain.SqlServer.MDS/SqlServer/CommandBuilders/ISupportsQueryHints.cs
@@ -1,0 +1,6 @@
+namespace Tortuga.Chain.SqlServer.CommandBuilders;
+
+interface ISupportsQueryHints
+{
+	void AddQueryHint(string hint);
+}