+++ Tortuga.Chain/Tortuga.Chain.SqlServer.MDS/SqlServer/CommandBuilders/ISupportsTableHints.cs
@@ -4,8 +4,3 @@ interface ISupportsTableHints
 {
 	void AddTableHint(string hint);
 }
-
-interface ISupportsQueryHints
-{
-	void AddQueryHint(string hint);
-}